feat: create @karmaniverous/jeeves-runner-core package (#73)
Extract runnerConfigSchema and RunnerConfig type into a new shared core
package. Service config.ts now re-exports from core.
