+++ packages/core/package.json
@@ -1,0 +1,86 @@
+{
+  "name": "@karmaniverous/jeeves-runner-core",
+  "version": "0.1.0",
+  "author": "Jason Williscroft",
+  "description": "Shared configuration schema and types for jeeves-runner packages.",
+  "license": "BSD-3-Clause",
+  "type": "module",
+  "files": [
+    "dist"
+  ],
+  "publishConfig": {
+    "access": "public"
+  },
+  "exports": {
+    ".": {
+      "import": {
+        "types": "./dist/src/index.d.ts",
+        "default": "./dist/src/index.js"
+      }
+    }
+  },
+  "types": "dist/src/index.d.ts",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/karmaniverous/jeeves-runner.git",
+    "directory": "packages/core"
+  },
+  "bugs": {
+    "url": "https://github.com/karmaniverous/jeeves-runner/issues"
+  },
+  "homepage": "https://github.com/karmaniverous/jeeves-runner#readme",
+  "engines": {
+    "node": ">=22"
+  },
+  "dependencies": {
+    "@karmaniverous/jeeves": "^0.5.10",
+    "zod": "^4.4.3"
+  },
+  "devDependencies": {
+    "@dotenvx/dotenvx": "^1.65.0",
+    "release-it": "^19.2.4",
+    "vitest": "^4.1.6"
+  },
+  "scripts": {
+    "prebuild": "npx rimraf dist",
+    "build": "tsc",
+    "lint": "eslint .",
+    "lint:fix": "eslint --fix .",
+    "test": "vitest run --passWithNoTests",
+    "typecheck": "tsc --noEmit",
+    "changelog": "git-cliff -o CHANGELOG.md",
+    "release": "dotenvx run -f .env.local -- release-it",
+    "release:pre": "dotenvx run -f .env.local -- release-it --no-git.requireBranch --github.prerelease --preRelease"
+  },
+  "release-it": {
+    "git": {
+      "changelog": "npx git-cliff --unreleased --strip header",
+      "commitMessage": "chore: release @karmaniverous/jeeves-runner-core v${version}",
+      "tagName": "core/${version}",
+      "requireBranch": "main"
+    },
+    "github": {
+      "release": true
+    },
+    "hooks": {
+      "after:init": [
+        "npm run lint",
+        "npm run typecheck",
+        "npm run test",
+        "npm run build"
+      ],
+      "before:npm:release": [
+        "npx git-cliff -o CHANGELOG.md",
+        "git add -A"
+      ],
+      "after:release": [
+        "git switch -c release/core/${version}",
+        "git push -u origin release/core/${version}",
+        "git switch ${branchName}"
+      ]
+    },
+    "npm": {
+      "publish": true
+    }
+  }
+}
