+++ packages/core/src/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Shared configuration schema and types for jeeves-runner packages.
+ *
+ * @packageDocumentation
+ */
+
+export { type RunnerConfig, runnerConfigSchema } from './config.js';
