+++ packages/service/package.json
@@ -47,6 +47,7 @@
   },
   "dependencies": {
     "@karmaniverous/jeeves": "^0.5.9",
+    "@karmaniverous/jeeves-runner-core": "*",
     "@karmaniverous/rrstack": "^0.17.1",
     "commander": "^14.0.3",
     "croner": "^10.0.1",
